backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
sqlparse==0.5.5
tzdata==2025.3
